requirements.txt
Django==4.1.7
djangorestframework==3.14.0
djangorestframework_simplejwt==5.0.0
rest_framework_simplejwt==0.0.1
